refactoring, more work on terminal
